webserver updates

+++ socketManager.ino
@@ -52,7 +52,7 @@ void onWsEvent(AsyncWebSocket* server, AsyncWebSocketClient* client, AwsEventTyp
             JsonArray value = jsonBuffer.createNestedArray("value");
             JsonObject key = value.createNestedObject();
             key["battery"] = random(100);
-            key["memory"] = (ESP.getFreeHeap()/ESP.getHeapSize())*100;
+            key["memory"] = long((ESP.getFreeHeap()/ESP.getHeapSize())*100);
             key["storage"] = random(100);
 
             serializeJson(jsonBuffer, reply);
